refactor: 인기 테마 조회 책임 theme으로 수정
