+++ src/main/java/roomescape/theme/response/RankTheme.java
@@ -1,4 +1,4 @@
-package roomescape.rank.response;
+package roomescape.theme.response;
 
 public record RankTheme(String name,String description,String thumbnail) {
 }
