+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 
 import roomescape.theme.domain.Theme;
 import roomescape.theme.repository.ThemeDao;
+import roomescape.theme.response.RankTheme;
 
 @Service
 public class ThemeService {
@@ -26,5 +27,9 @@ public class ThemeService {
 
     public void delete(long themeId){
         themeDao.delete(themeId);
+    }
+
+    public List<RankTheme> getRank() {
+        return themeDao.getRank();
     }
 }
